Broke up JournalMessage into smaller types.
The JournalMessage type worked before okay, but was pretty huge
and unwieldy.  Now it is separated into several types.  The only
drawback is that now the caller has to do type assertions on the
Message field in Journal to determine which type was sent.  However
the benefit is that after the assertion is handled the caller now
has a better idea of exactly what data they have rather than guessing
via the previous huge JournalMessage type.